refactor: blob actor files
