+++ fendermint/actors/blob_reader/Cargo.toml
@@ -32,8 +32,7 @@ recall_ipld = { path = "../../../recall/ipld" }
 fil_actors_evm_shared = { workspace = true }
 fil_actors_runtime = { workspace = true, features = ["test_utils"] }
 hex-literal = { workspace = true }
-iroh-base = { workspace = true, features = ["key"] }
-rand = { workspace = true }
+fendermint_actor_blobs_testing = { path = "../blobs/testing" }
 
 [features]
 fil-actor = ["fil_actors_runtime/fil-actor"]
