+++ fendermint/actors/blobs/src/actor/user.rs
@@ -1,0 +1,1124 @@
+// Copyright 2025 Recall Contributors
+// Copyright 2021-2023 Protocol Labs
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use fendermint_actor_blobs_shared::params::{
+    AddBlobParams, ApproveCreditParams, BuyCreditParams, DeleteBlobParams, GetAccountParams,
+    GetBlobParams, GetCreditApprovalParams, OverwriteBlobParams, RevokeCreditParams,
+    SetSponsorParams,
+};
+use fendermint_actor_blobs_shared::state::{
+    AccountInfo, Blob, Credit, CreditApproval, Subscription,
+};
+use fendermint_actor_machine::{
+    caller::{Caller, CallerOption},
+    events::emit_evm_event,
+    util::{to_delegated_address, token_to_biguint},
+};
+use fendermint_actor_recall_config_shared::get_config;
+use fil_actors_runtime::{extract_send_result, runtime::Runtime, ActorError};
+use fvm_shared::{econ::TokenAmount, METHOD_SEND};
+use num_traits::Zero;
+use recall_sol_facade::{
+    blobs::{blob_added, blob_deleted},
+    credit::{credit_approved, credit_purchased, credit_revoked},
+    gas::{gas_sponsor_set, gas_sponsor_unset},
+};
+
+use crate::actor::{delete_from_disc, BlobsActor};
+use crate::caller::DelegationOptions;
+use crate::state::blobs::{AddBlobStateParams, DeleteBlobStateParams};
+use crate::State;
+
+impl BlobsActor {
+    /// Buy credit with token.
+    ///
+    /// The `to` address must be delegated (only delegated addresses can own credit).
+    pub fn buy_credit(
+        rt: &impl Runtime,
+        params: BuyCreditParams,
+    ) -> Result<AccountInfo, ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let caller = Caller::new_delegated(rt, params.0, None, CallerOption::Auth)?;
+        let config = get_config(rt)?;
+
+        let mut credit_amount = Credit::zero();
+        let account = rt.transaction(|st: &mut State, rt| {
+            let pre_buy = st.credit_sold.clone();
+            let account = st.buy_credit(
+                rt.store(),
+                &config,
+                caller.state_address(),
+                rt.message().value_received(),
+                rt.curr_epoch(),
+            )?;
+            credit_amount = &st.credit_sold - &pre_buy;
+            Ok(account)
+        })?;
+
+        emit_evm_event(
+            rt,
+            credit_purchased(
+                caller.event_address(),
+                token_to_biguint(Some(credit_amount)),
+            ),
+        )?;
+
+        AccountInfo::from(account, rt)
+    }
+
+    /// Approve credit and gas usage from one account to another.
+    ///
+    /// The `from` address must be delegated (only delegated addresses can own credit).
+    /// The `from` address must be the message origin or caller.
+    /// The `to` address must be delegated (only delegated addresses can use credit).
+    /// The `to` address will be created if it does not exist.
+    /// TODO: Remove the `caller_allowlist` parameter.
+    pub fn approve_credit(
+        rt: &impl Runtime,
+        params: ApproveCreditParams,
+    ) -> Result<CreditApproval, ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let from_caller = Caller::new_delegated(rt, params.from, None, CallerOption::Auth)?;
+        let to_caller = Caller::new_delegated(rt, params.to, None, CallerOption::Create)?;
+        let config = get_config(rt)?;
+
+        let approval = rt.transaction(|st: &mut State, rt| {
+            let approval = st.approve_credit(
+                &config,
+                rt.store(),
+                from_caller.state_address(),
+                to_caller.state_address(),
+                DelegationOptions {
+                    credit_limit: params.credit_limit,
+                    gas_fee_limit: params.gas_fee_limit,
+                    ttl: params.ttl,
+                },
+                rt.curr_epoch(),
+            );
+
+            // For convenience, set the approvee's sponsor to the approver if it was created
+            if to_caller.created() {
+                st.set_account_sponsor(
+                    &config,
+                    rt.store(),
+                    to_caller.state_address(),
+                    Some(from_caller.state_address()),
+                    rt.curr_epoch(),
+                )?;
+            }
+            approval
+        })?;
+
+        let event_credit_limit = token_to_biguint(approval.credit_limit.clone());
+        let event_gas_fee_limit = token_to_biguint(approval.gas_allowance_limit.clone());
+        let event_expiry = approval.expiry.unwrap_or_default() as u64;
+        emit_evm_event(
+            rt,
+            credit_approved(
+                from_caller.event_address(),
+                to_caller.event_address(),
+                event_credit_limit,
+                event_gas_fee_limit,
+                event_expiry,
+            ),
+        )?;
+
+        Ok(approval)
+    }
+
+    /// Revoke credit and gas usage from one account to another.
+    ///
+    /// The `from` address must be delegated (only delegated addresses can own credit).
+    /// The `from` address must be the message origin or caller.
+    /// The `to` address must be delegated (only delegated addresses can use credit).
+    pub fn revoke_credit(rt: &impl Runtime, params: RevokeCreditParams) -> Result<(), ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let from_caller = Caller::new_delegated(rt, params.from, None, CallerOption::Auth)?;
+        let to_caller = Caller::new_delegated(rt, params.to, None, CallerOption::None)?;
+
+        rt.transaction(|st: &mut State, rt| {
+            st.revoke_credit(
+                rt.store(),
+                from_caller.state_address(),
+                to_caller.state_address(),
+            )
+        })?;
+
+        emit_evm_event(
+            rt,
+            credit_revoked(from_caller.event_address(), to_caller.event_address()),
+        )?;
+
+        Ok(())
+    }
+
+    /// Sets or unsets a default credit and gas sponsor from one account to another.
+    ///
+    /// If `sponsor` does not exist, the default sponsor is unset.
+    /// The `from` address must be delegated (only delegated addresses can use credit).
+    /// The `from` address must be the message origin or caller.
+    /// The `sponsor` address must be delegated (only delegated addresses can own credit).
+    pub fn set_account_sponsor(
+        rt: &impl Runtime,
+        params: SetSponsorParams,
+    ) -> Result<(), ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let caller = Caller::new_delegated(rt, params.from, params.sponsor, CallerOption::Auth)?;
+        let config = get_config(rt)?;
+
+        rt.transaction(|st: &mut State, rt| {
+            st.set_account_sponsor(
+                &config,
+                rt.store(),
+                caller.state_address(),
+                caller.sponsor_state_address(),
+                rt.curr_epoch(),
+            )
+        })?;
+
+        if let Some(sponsor) = caller.sponsor_address() {
+            emit_evm_event(rt, gas_sponsor_set(sponsor))?;
+        } else {
+            emit_evm_event(rt, gas_sponsor_unset())?;
+        }
+
+        Ok(())
+    }
+
+    /// Returns the account for an address.
+    pub fn get_account(
+        rt: &impl Runtime,
+        params: GetAccountParams,
+    ) -> Result<Option<AccountInfo>, ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let caller = Caller::new(rt, params.0, None, CallerOption::None)?;
+
+        let account = rt
+            .state::<State>()?
+            .get_account(rt.store(), caller.state_address())?
+            .map(|mut account| {
+                // Resolve the credit sponsor
+                account.credit_sponsor = account
+                    .credit_sponsor
+                    .map(|sponsor| to_delegated_address(rt, sponsor))
+                    .transpose()?;
+
+                AccountInfo::from(account, rt)
+            });
+
+        account.transpose()
+    }
+
+    /// Returns the credit approval from one account to another if it exists.
+    pub fn get_credit_approval(
+        rt: &impl Runtime,
+        params: GetCreditApprovalParams,
+    ) -> Result<Option<CreditApproval>, ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let from_caller = Caller::new(rt, params.from, None, CallerOption::None)?;
+        let to_caller = Caller::new(rt, params.to, None, CallerOption::None)?;
+
+        let approval = rt.state::<State>()?.get_credit_approval(
+            rt.store(),
+            from_caller.state_address(),
+            to_caller.state_address(),
+        )?;
+
+        Ok(approval)
+    }
+
+    /// Adds or updates a blob subscription.
+    ///
+    /// The subscriber will only need credits for blobs that are not already covered by one of
+    /// their existing subscriptions.
+    ///
+    /// The `sponsor` will be the subscriber (the account responsible for payment), if it exists
+    /// and there is an approval from `sponsor` to `from`.    
+    ///
+    /// The `from` address must be delegated (only delegated addresses can use credit).
+    /// The `sponsor` address must be delegated (only delegated addresses can use credit).
+    pub fn add_blob(rt: &impl Runtime, params: AddBlobParams) -> Result<Subscription, ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let caller = Caller::new_delegated(rt, params.from, params.sponsor, CallerOption::Auth)?;
+        let token_amount = rt.message().value_received();
+        let config = get_config(rt)?;
+
+        let mut capacity_used = 0;
+        let (sub, token_rebate) = rt.transaction(|st: &mut State, rt| {
+            let initial_capacity_used = st.capacity_used;
+            let res = st.add_blob(
+                rt.store(),
+                &config,
+                caller.state_address(),
+                caller.sponsor_state_address(),
+                AddBlobStateParams::from_actor_params(
+                    params.clone(),
+                    rt.curr_epoch(),
+                    token_amount,
+                ),
+            )?;
+            capacity_used = st.capacity_used - initial_capacity_used;
+            Ok(res)
+        })?;
+
+        // Send back unspent tokens
+        if !token_rebate.is_zero() {
+            extract_send_result(rt.send_simple(
+                &caller.state_address(),
+                METHOD_SEND,
+                None,
+                token_rebate,
+            ))?;
+        }
+
+        emit_evm_event(
+            rt,
+            blob_added(
+                caller.event_address(),
+                &params.hash.0,
+                params.size,
+                sub.expiry as u64,
+                capacity_used,
+            ),
+        )?;
+
+        Ok(sub)
+    }
+
+    /// Returns a blob by hash if it exists.
+    pub fn get_blob(rt: &impl Runtime, params: GetBlobParams) -> Result<Option<Blob>, ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+        let blob = rt.state::<State>()?.get_blob(rt.store(), params.0)?;
+        Ok(blob)
+    }
+
+    /// Deletes a blob subscription.
+    ///
+    /// The `sponsor` will be the subscriber (the account responsible for payment), if it exists
+    /// and there is an approval from `sponsor` to `from`.    
+    ///
+    /// The `from` address must be delegated (only delegated addresses can use credit).
+    /// The `sponsor` address must be delegated (only delegated addresses can use credit).
+    pub fn delete_blob(rt: &impl Runtime, params: DeleteBlobParams) -> Result<(), ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let caller = Caller::new_delegated(rt, params.from, params.sponsor, CallerOption::Auth)?;
+
+        let mut capacity_released = 0;
+        let (delete, size) = rt.transaction(|st: &mut State, rt| {
+            let initial_capacity_used = st.capacity_used;
+            let res = st.delete_blob(
+                rt.store(),
+                caller.state_address(),
+                caller.sponsor_state_address(),
+                DeleteBlobStateParams::from_actor_params(params.clone(), rt.curr_epoch()),
+            )?;
+            capacity_released = initial_capacity_used - st.capacity_used;
+            Ok(res)
+        })?;
+
+        if delete {
+            delete_from_disc(params.hash)?;
+        }
+
+        emit_evm_event(
+            rt,
+            blob_deleted(
+                caller.event_address(),
+                &params.hash.0,
+                size,
+                capacity_released,
+            ),
+        )?;
+
+        Ok(())
+    }
+
+    /// Deletes a blob subscription and adds another in a sinlge call.
+    ///
+    /// This method is more efficient than two separate calls to `delete_blob` and `add_blob`,
+    /// and is useful for some blob workflows like a replacing a key in a bucket actor.
+    ///
+    /// The `sponsor` will be the subscriber (the account responsible for payment), if it exists
+    /// and there is an approval from `sponsor` to `from`.    
+    ///
+    /// The `from` address must be delegated (only delegated addresses can use credit).
+    /// The `sponsor` address must be delegated (only delegated addresses can use credit).
+    pub fn overwrite_blob(
+        rt: &impl Runtime,
+        params: OverwriteBlobParams,
+    ) -> Result<Subscription, ActorError> {
+        rt.validate_immediate_caller_accept_any()?;
+
+        let caller =
+            Caller::new_delegated(rt, params.add.from, params.add.sponsor, CallerOption::Auth)?;
+        let config = get_config(rt)?;
+
+        // Determine if we need to delete an existing blob before adding the new one
+        let overwrite = params.old_hash != params.add.hash;
+
+        let add_hash = params.add.hash;
+        let add_size = params.add.size;
+        let mut capacity_released = 0;
+        let mut capacity_used = 0;
+
+        // To ensure atomicity, we combine the two independent calls into a single transaction.
+        let (delete, delete_size, sub) = rt.transaction(|st: &mut State, rt| {
+            let add_params = params.add;
+
+            let initial_capacity_used = st.capacity_used;
+            let (delete, delete_size) = if overwrite {
+                st.delete_blob(
+                    rt.store(),
+                    caller.state_address(),
+                    caller.sponsor_state_address(),
+                    DeleteBlobStateParams {
+                        hash: params.old_hash,
+                        id: add_params.id.clone(),
+                        epoch: rt.curr_epoch(),
+                    },
+                )?
+            } else {
+                (false, 0)
+            };
+            capacity_released = initial_capacity_used - st.capacity_used;
+
+            let initial_capacity_used = st.capacity_used;
+            let (subscription, _) = st.add_blob(
+                rt.store(),
+                &config,
+                caller.state_address(),
+                caller.sponsor_state_address(),
+                AddBlobStateParams::from_actor_params(
+                    add_params,
+                    rt.curr_epoch(),
+                    TokenAmount::zero(),
+                ),
+            )?;
+            capacity_used = st.capacity_used - initial_capacity_used;
+
+            Ok((delete, delete_size, subscription))
+        })?;
+
+        if delete {
+            delete_from_disc(params.old_hash)?;
+        }
+
+        if overwrite {
+            emit_evm_event(
+                rt,
+                blob_deleted(
+                    caller.event_address(),
+                    &params.old_hash.0,
+                    delete_size,
+                    capacity_released,
+                ),
+            )?;
+        }
+        emit_evm_event(
+            rt,
+            blob_added(
+                caller.event_address(),
+                &add_hash.0,
+                add_size,
+                sub.expiry as u64,
+                capacity_used,
+            ),
+        )?;
+
+        Ok(sub)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::testing::{
+        construct_and_verify, expect_emitted_add_event, expect_emitted_approve_event,
+        expect_emitted_purchase_event, expect_emitted_revoke_event, expect_get_config,
+    };
+    use fendermint_actor_blobs_shared::state::SubscriptionId;
+    use fendermint_actor_blobs_shared::Method;
+    use fendermint_actor_blobs_testing::{new_hash, new_pk, setup_logs};
+    use fil_actors_evm_shared::address::EthAddress;
+    use fil_actors_runtime::test_utils::{ETHACCOUNT_ACTOR_CODE_ID, EVM_ACTOR_CODE_ID};
+    use fvm_ipld_encoding::ipld_block::IpldBlock;
+    use fvm_shared::{address::Address, bigint::BigInt, clock::ChainEpoch, error::ExitCode};
+
+    #[test]
+    fn test_buy_credit() {
+        setup_logs();
+        let rt = construct_and_verify();
+
+        // TODO(bcalza): Choose a rate different than default
+        let token_credit_rate = BigInt::from(1000000000000000000u64);
+
+        let id_addr = Address::new_id(110);
+        let eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000000"
+        ));
+        let f4_eth_addr = Address::new_delegated(10, &eth_addr.0).unwrap();
+
+        rt.set_delegated_address(id_addr.id().unwrap(), f4_eth_addr);
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, id_addr);
+        rt.set_origin(id_addr);
+
+        let tokens = 1;
+        let mut expected_credits =
+            Credit::from_atto(1000000000000000000u64 * tokens * &token_credit_rate);
+        let mut expected_gas_allowance = TokenAmount::from_whole(tokens);
+        rt.set_received(TokenAmount::from_whole(tokens));
+        rt.expect_validate_caller_any();
+        let fund_params = BuyCreditParams(f4_eth_addr);
+        expect_get_config(&rt);
+        expect_emitted_purchase_event(&rt, &fund_params, expected_credits.clone());
+        let result = rt
+            .call::<BlobsActor>(
+                Method::BuyCredit as u64,
+                IpldBlock::serialize_cbor(&fund_params).unwrap(),
+            )
+            .unwrap()
+            .unwrap()
+            .deserialize::<AccountInfo>()
+            .unwrap();
+        assert_eq!(result.credit_free, expected_credits);
+        assert_eq!(result.gas_allowance, expected_gas_allowance);
+        rt.verify();
+
+        let additional_credits = Credit::from_atto(1000000000u64 * tokens * &token_credit_rate);
+        expected_credits += &additional_credits;
+        expected_gas_allowance += TokenAmount::from_nano(tokens);
+        rt.set_received(TokenAmount::from_nano(tokens));
+        rt.expect_validate_caller_any();
+        let fund_params = BuyCreditParams(f4_eth_addr);
+        expect_get_config(&rt);
+        expect_emitted_purchase_event(&rt, &fund_params, additional_credits);
+        let result = rt
+            .call::<BlobsActor>(
+                Method::BuyCredit as u64,
+                IpldBlock::serialize_cbor(&fund_params).unwrap(),
+            )
+            .unwrap()
+            .unwrap()
+            .deserialize::<AccountInfo>()
+            .unwrap();
+        assert_eq!(result.credit_free, expected_credits);
+        assert_eq!(result.gas_allowance, expected_gas_allowance);
+        rt.verify();
+
+        let additional_credits = Credit::from_atto(tokens * &token_credit_rate);
+        expected_credits += &additional_credits;
+        expected_gas_allowance += TokenAmount::from_atto(tokens);
+        rt.set_received(TokenAmount::from_atto(tokens));
+        rt.expect_validate_caller_any();
+        let fund_params = BuyCreditParams(f4_eth_addr);
+        expect_get_config(&rt);
+        expect_emitted_purchase_event(&rt, &fund_params, additional_credits);
+        let result = rt
+            .call::<BlobsActor>(
+                Method::BuyCredit as u64,
+                IpldBlock::serialize_cbor(&fund_params).unwrap(),
+            )
+            .unwrap()
+            .unwrap()
+            .deserialize::<AccountInfo>()
+            .unwrap();
+        assert_eq!(result.credit_free, expected_credits);
+        assert_eq!(result.gas_allowance, expected_gas_allowance);
+        rt.verify();
+    }
+
+    #[test]
+    fn test_approve_credit() {
+        setup_logs();
+        let rt = construct_and_verify();
+
+        // Credit owner
+        let owner_id_addr = Address::new_id(110);
+        let owner_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000000"
+        ));
+        let owner_f4_eth_addr = Address::new_delegated(10, &owner_eth_addr.0).unwrap();
+        rt.set_delegated_address(owner_id_addr.id().unwrap(), owner_f4_eth_addr);
+
+        // Credit receiver
+        let to_id_addr = Address::new_id(111);
+        let to_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000001"
+        ));
+        let to_f4_eth_addr = Address::new_delegated(10, &to_eth_addr.0).unwrap();
+        rt.set_delegated_address(to_id_addr.id().unwrap(), to_f4_eth_addr);
+        rt.set_address_actor_type(to_id_addr, *ETHACCOUNT_ACTOR_CODE_ID);
+
+        // Proxy EVM contract on behalf of the credit owner
+        let proxy_id_addr = Address::new_id(112);
+        let proxy_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000002"
+        ));
+        let proxy_f4_eth_addr = Address::new_delegated(10, &proxy_eth_addr.0).unwrap();
+        rt.set_delegated_address(proxy_id_addr.id().unwrap(), proxy_f4_eth_addr);
+        rt.set_address_actor_type(proxy_id_addr, *EVM_ACTOR_CODE_ID);
+
+        // Caller/origin is the same as from (i.e., the standard case)
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, owner_id_addr);
+        rt.set_origin(owner_id_addr);
+        rt.expect_validate_caller_any();
+        expect_get_config(&rt);
+        let approve_params = ApproveCreditParams {
+            from: owner_id_addr,
+            to: to_id_addr,
+            caller_allowlist: None,
+            credit_limit: None,
+            gas_fee_limit: None,
+            ttl: None,
+        };
+        expect_emitted_approve_event(
+            &rt,
+            owner_f4_eth_addr,
+            to_f4_eth_addr,
+            approve_params.credit_limit.clone(),
+            approve_params.gas_fee_limit.clone(),
+            0,
+        );
+        let result = rt.call::<BlobsActor>(
+            Method::ApproveCredit as u64,
+            IpldBlock::serialize_cbor(&approve_params).unwrap(),
+        );
+        assert!(result.is_ok());
+        rt.verify();
+
+        // Proxy caller (caller mismatch with from, but is correct origin)
+        rt.set_caller(*EVM_ACTOR_CODE_ID, proxy_id_addr);
+        rt.set_origin(owner_id_addr);
+        rt.expect_validate_caller_any();
+        expect_get_config(&rt);
+        let approve_params = ApproveCreditParams {
+            from: owner_id_addr,
+            to: to_id_addr,
+            caller_allowlist: None,
+            credit_limit: None,
+            gas_fee_limit: None,
+            ttl: None,
+        };
+        expect_emitted_approve_event(
+            &rt,
+            owner_f4_eth_addr,
+            to_f4_eth_addr,
+            approve_params.credit_limit.clone(),
+            approve_params.gas_fee_limit.clone(),
+            0,
+        );
+        let result = rt.call::<BlobsActor>(
+            Method::ApproveCredit as u64,
+            IpldBlock::serialize_cbor(&approve_params).unwrap(),
+        );
+        assert!(result.is_ok());
+        rt.verify();
+
+        // Caller/origin mismatch with from
+        rt.set_caller(*EVM_ACTOR_CODE_ID, proxy_id_addr);
+        rt.set_origin(owner_id_addr);
+        rt.expect_validate_caller_any();
+        let approve_params = ApproveCreditParams {
+            from: to_id_addr, // mismatch
+            to: to_id_addr,
+            caller_allowlist: None,
+            credit_limit: None,
+            gas_fee_limit: None,
+            ttl: None,
+        };
+        let result = rt.call::<BlobsActor>(
+            Method::ApproveCredit as u64,
+            IpldBlock::serialize_cbor(&approve_params).unwrap(),
+        );
+        let expected_return = Err(ActorError::illegal_argument(format!(
+            "address {} does not match origin or caller",
+            to_id_addr
+        )));
+        assert_eq!(result, expected_return);
+        rt.verify();
+    }
+
+    #[test]
+    fn test_approve_credit_to_new_account() {
+        setup_logs();
+        let rt = construct_and_verify();
+
+        // Credit owner
+        let owner_id_addr = Address::new_id(110);
+        let owner_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000000"
+        ));
+        let owner_f4_eth_addr = Address::new_delegated(10, &owner_eth_addr.0).unwrap();
+        rt.set_delegated_address(owner_id_addr.id().unwrap(), owner_f4_eth_addr);
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, owner_id_addr);
+        rt.set_origin(owner_id_addr);
+
+        // Use a new receiver that doesn't exist in the FVM
+        let receiver_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000001"
+        ));
+        let receiver_f4_eth_addr = Address::new_delegated(10, &receiver_eth_addr.0).unwrap();
+
+        rt.expect_validate_caller_any();
+        rt.expect_send_simple(
+            receiver_f4_eth_addr,
+            METHOD_SEND,
+            None,
+            TokenAmount::zero(),
+            None,
+            ExitCode::OK,
+        );
+        let approve_params = ApproveCreditParams {
+            from: owner_f4_eth_addr,
+            to: receiver_f4_eth_addr, // Use the external address to force the ID lookup to fail
+            caller_allowlist: None,
+            credit_limit: None,
+            gas_fee_limit: None,
+            ttl: None,
+        };
+        let result = rt.call::<BlobsActor>(
+            Method::ApproveCredit as u64,
+            IpldBlock::serialize_cbor(&approve_params).unwrap(),
+        );
+        // This test should pass, but in the mock runtime, sending token to an address does not
+        // create the actor, like it does in the real FVM runtime.
+        // The result is that the second call to to_id_address in the approve_credit method still
+        // fails after the call to send with a "not found" error.
+        // However, we are able to test that the call to send did happen using
+        // rt.expect_send_simple above.
+        assert!(result.is_err());
+        assert_eq!(result.unwrap_err().exit_code(), ExitCode::USR_NOT_FOUND);
+        rt.verify();
+    }
+
+    #[test]
+    fn test_revoke_credit() {
+        setup_logs();
+        let rt = construct_and_verify();
+
+        // Credit owner
+        let owner_id_addr = Address::new_id(110);
+        let owner_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000000"
+        ));
+        let owner_f4_eth_addr = Address::new_delegated(10, &owner_eth_addr.0).unwrap();
+        rt.set_delegated_address(owner_id_addr.id().unwrap(), owner_f4_eth_addr);
+
+        // Credit receiver
+        let to_id_addr = Address::new_id(111);
+        let to_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000001"
+        ));
+        let to_f4_eth_addr = Address::new_delegated(10, &to_eth_addr.0).unwrap();
+        rt.set_delegated_address(to_id_addr.id().unwrap(), to_f4_eth_addr);
+        rt.set_address_actor_type(to_id_addr, *ETHACCOUNT_ACTOR_CODE_ID);
+
+        // Proxy EVM contract on behalf of the credit owner
+        let proxy_id_addr = Address::new_id(112);
+        let proxy_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000002"
+        ));
+        let proxy_f4_eth_addr = Address::new_delegated(10, &proxy_eth_addr.0).unwrap();
+        rt.set_delegated_address(proxy_id_addr.id().unwrap(), proxy_f4_eth_addr);
+        rt.set_address_actor_type(proxy_id_addr, *EVM_ACTOR_CODE_ID);
+
+        // Set up the approval to revoke
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, owner_id_addr);
+        rt.set_origin(owner_id_addr);
+        rt.expect_validate_caller_any();
+        expect_get_config(&rt);
+        let approve_params = ApproveCreditParams {
+            from: owner_id_addr,
+            to: to_id_addr,
+            caller_allowlist: None,
+            credit_limit: None,
+            gas_fee_limit: None,
+            ttl: None,
+        };
+        expect_emitted_approve_event(
+            &rt,
+            owner_f4_eth_addr,
+            to_f4_eth_addr,
+            approve_params.credit_limit.clone(),
+            approve_params.gas_fee_limit.clone(),
+            0,
+        );
+        let result = rt.call::<BlobsActor>(
+            Method::ApproveCredit as u64,
+            IpldBlock::serialize_cbor(&approve_params).unwrap(),
+        );
+        assert!(result.is_ok());
+        rt.verify();
+
+        // Caller/origin is the same as from (i.e., the standard case)
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, owner_id_addr);
+        rt.set_origin(owner_id_addr);
+        rt.expect_validate_caller_any();
+        let revoke_params = RevokeCreditParams {
+            from: owner_id_addr,
+            to: to_id_addr,
+            for_caller: None,
+        };
+        expect_emitted_revoke_event(&rt, owner_f4_eth_addr, to_f4_eth_addr);
+        let result = rt.call::<BlobsActor>(
+            Method::RevokeCredit as u64,
+            IpldBlock::serialize_cbor(&revoke_params).unwrap(),
+        );
+        assert!(result.is_ok());
+        rt.verify();
+
+        // Proxy caller (caller mismatch with from, but is correct origin)
+        rt.set_caller(*EVM_ACTOR_CODE_ID, proxy_id_addr);
+        rt.set_origin(owner_id_addr);
+        rt.expect_validate_caller_any();
+        let revoke_params = RevokeCreditParams {
+            from: owner_id_addr,
+            to: to_id_addr,
+            for_caller: None,
+        };
+        let result = rt.call::<BlobsActor>(
+            Method::RevokeCredit as u64,
+            IpldBlock::serialize_cbor(&revoke_params).unwrap(),
+        );
+        // This should be a state error, not from the actor API
+        assert!(result.is_err());
+        assert!(result.err().unwrap().msg().contains("not found"),);
+        rt.verify();
+
+        // Caller/origin mismatch with from
+        rt.set_caller(*EVM_ACTOR_CODE_ID, proxy_id_addr);
+        rt.set_origin(owner_id_addr);
+        rt.expect_validate_caller_any();
+        let revoke_params = RevokeCreditParams {
+            from: to_id_addr, // mismatch
+            to: to_id_addr,
+            for_caller: None,
+        };
+        let result = rt.call::<BlobsActor>(
+            Method::RevokeCredit as u64,
+            IpldBlock::serialize_cbor(&revoke_params).unwrap(),
+        );
+        let expected_return = Err(ActorError::illegal_argument(format!(
+            "address {} does not match origin or caller",
+            to_id_addr
+        )));
+        assert_eq!(result, expected_return);
+        rt.verify();
+    }
+
+    #[test]
+    fn test_add_blob() {
+        setup_logs();
+        let rt = construct_and_verify();
+
+        let token_credit_rate = BigInt::from(1000000000000000000u64);
+
+        let id_addr = Address::new_id(110);
+        let eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000000"
+        ));
+        let f4_eth_addr = Address::new_delegated(10, &eth_addr.0).unwrap();
+
+        rt.set_delegated_address(id_addr.id().unwrap(), f4_eth_addr);
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, id_addr);
+        rt.set_origin(id_addr);
+        rt.set_epoch(ChainEpoch::from(0));
+
+        // Try without first funding
+        rt.expect_validate_caller_any();
+        let hash = new_hash(1024);
+        let add_params = AddBlobParams {
+            from: id_addr,
+            sponsor: None,
+            source: new_pk(),
+            hash: hash.0,
+            metadata_hash: new_hash(1024).0,
+            id: SubscriptionId::default(),
+            size: hash.1,
+            ttl: Some(3600),
+        };
+        expect_get_config(&rt);
+        let result = rt.call::<BlobsActor>(
+            Method::AddBlob as u64,
+            IpldBlock::serialize_cbor(&add_params).unwrap(),
+        );
+        assert!(result.is_err());
+        rt.verify();
+
+        // Fund an account
+        let tokens = 1;
+        let received = TokenAmount::from_whole(tokens);
+        let expected_credits =
+            Credit::from_atto(1000000000000000000u64 * tokens * &token_credit_rate);
+        rt.set_received(received.clone());
+        rt.expect_validate_caller_any();
+        let fund_params = BuyCreditParams(f4_eth_addr);
+        expect_get_config(&rt);
+        expect_emitted_purchase_event(&rt, &fund_params, expected_credits);
+        let result = rt.call::<BlobsActor>(
+            Method::BuyCredit as u64,
+            IpldBlock::serialize_cbor(&fund_params).unwrap(),
+        );
+        assert!(result.is_ok());
+        rt.verify();
+
+        // Try with sufficient balance
+        rt.set_received(TokenAmount::zero());
+        rt.set_epoch(ChainEpoch::from(5));
+        rt.expect_validate_caller_any();
+        expect_get_config(&rt);
+        expect_emitted_add_event(&rt, 5, &add_params, f4_eth_addr, add_params.size);
+        let subscription = rt
+            .call::<BlobsActor>(
+                Method::AddBlob as u64,
+                IpldBlock::serialize_cbor(&add_params).unwrap(),
+            )
+            .unwrap()
+            .unwrap()
+            .deserialize::<Subscription>()
+            .unwrap();
+        assert_eq!(subscription.added, 5);
+        assert_eq!(subscription.expiry, 3605);
+        assert_eq!(subscription.delegate, None);
+        rt.verify();
+    }
+
+    #[test]
+    fn test_add_blob_inline_buy() {
+        setup_logs();
+        let rt = construct_and_verify();
+
+        let id_addr = Address::new_id(110);
+        let eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000000"
+        ));
+        let f4_eth_addr = Address::new_delegated(10, &eth_addr.0).unwrap();
+
+        rt.set_delegated_address(id_addr.id().unwrap(), f4_eth_addr);
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, id_addr);
+        rt.set_origin(id_addr);
+        rt.set_epoch(ChainEpoch::from(0));
+
+        // Try sending a lot
+        rt.expect_validate_caller_any();
+        let hash = new_hash(1024);
+        let add_params = AddBlobParams {
+            from: id_addr,
+            sponsor: None,
+            source: new_pk(),
+            hash: hash.0,
+            metadata_hash: new_hash(1024).0,
+            id: SubscriptionId::default(),
+            size: hash.1,
+            ttl: Some(3600),
+        };
+        let tokens_sent = TokenAmount::from_whole(1);
+        rt.set_received(tokens_sent.clone());
+        rt.set_balance(tokens_sent.clone());
+        let tokens_required_atto = add_params.size * add_params.ttl.unwrap() as u64;
+        let expected_tokens_unspent = tokens_sent.atto() - tokens_required_atto;
+        expect_get_config(&rt);
+        expect_emitted_add_event(&rt, 0, &add_params, f4_eth_addr, add_params.size);
+        rt.expect_send_simple(
+            id_addr,
+            METHOD_SEND,
+            None,
+            TokenAmount::from_atto(expected_tokens_unspent),
+            None,
+            ExitCode::OK,
+        );
+        let result = rt.call::<BlobsActor>(
+            Method::AddBlob as u64,
+            IpldBlock::serialize_cbor(&add_params).unwrap(),
+        );
+        assert!(result.is_ok());
+        rt.verify();
+
+        // Try sending zero
+        rt.expect_validate_caller_any();
+        rt.set_received(TokenAmount::zero());
+        let hash = new_hash(1024);
+        let add_params = AddBlobParams {
+            from: id_addr,
+            sponsor: None,
+            hash: hash.0,
+            metadata_hash: new_hash(1024).0,
+            source: new_pk(),
+            id: SubscriptionId::default(),
+            size: hash.1,
+            ttl: Some(3600),
+        };
+        expect_get_config(&rt);
+        let response = rt.call::<BlobsActor>(
+            Method::AddBlob as u64,
+            IpldBlock::serialize_cbor(&add_params).unwrap(),
+        );
+        assert!(response.is_err());
+        rt.verify();
+
+        // Try sending exact amount
+        let tokens_required_atto = add_params.size * add_params.ttl.unwrap() as u64;
+        let tokens_sent = TokenAmount::from_atto(tokens_required_atto);
+        rt.set_received(tokens_sent.clone());
+        rt.expect_validate_caller_any();
+        let hash = new_hash(1024);
+        let add_params = AddBlobParams {
+            from: id_addr,
+            sponsor: None,
+            hash: hash.0,
+            metadata_hash: new_hash(1024).0,
+            source: new_pk(),
+            id: SubscriptionId::default(),
+            size: hash.1,
+            ttl: Some(3600),
+        };
+        expect_get_config(&rt);
+        expect_emitted_add_event(&rt, 0, &add_params, f4_eth_addr, add_params.size);
+        let result = rt.call::<BlobsActor>(
+            Method::AddBlob as u64,
+            IpldBlock::serialize_cbor(&add_params).unwrap(),
+        );
+        assert!(result.is_ok());
+        rt.verify();
+    }
+
+    #[test]
+    fn test_add_blob_with_sponsor() {
+        setup_logs();
+        let rt = construct_and_verify();
+
+        let token_credit_rate = BigInt::from(1000000000000000000u64);
+
+        // Credit sponsor
+        let sponsor_id_addr = Address::new_id(110);
+        let sponsor_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000000"
+        ));
+        let sponsor_f4_eth_addr = Address::new_delegated(10, &sponsor_eth_addr.0).unwrap();
+        rt.set_delegated_address(sponsor_id_addr.id().unwrap(), sponsor_f4_eth_addr);
+
+        // Credit spender
+        let spender_id_addr = Address::new_id(111);
+        let spender_eth_addr = EthAddress(hex_literal::hex!(
+            "CAFEB0BA00000000000000000000000000000001"
+        ));
+        let spender_f4_eth_addr = Address::new_delegated(10, &spender_eth_addr.0).unwrap();
+        rt.set_delegated_address(spender_id_addr.id().unwrap(), spender_f4_eth_addr);
+        rt.set_address_actor_type(spender_id_addr, *ETHACCOUNT_ACTOR_CODE_ID);
+
+        // Sponsor buys credit
+        let tokens = 1;
+        let received = TokenAmount::from_whole(tokens);
+        let expected_credits =
+            Credit::from_atto(1000000000000000000u64 * tokens * &token_credit_rate);
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, sponsor_id_addr);
+        rt.set_received(received);
+        rt.expect_validate_caller_any();
+        let fund_params = BuyCreditParams(sponsor_f4_eth_addr);
+        expect_get_config(&rt);
+        expect_emitted_purchase_event(&rt, &fund_params, expected_credits);
+        let response = rt.call::<BlobsActor>(
+            Method::BuyCredit as u64,
+            IpldBlock::serialize_cbor(&fund_params).unwrap(),
+        );
+        assert!(response.is_ok());
+        rt.verify();
+
+        // Sponsors approves credit
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, sponsor_id_addr);
+        rt.set_origin(sponsor_id_addr);
+        rt.expect_validate_caller_any();
+        expect_get_config(&rt);
+        let approve_params = ApproveCreditParams {
+            from: sponsor_id_addr,
+            to: spender_id_addr,
+            caller_allowlist: None,
+            credit_limit: None,
+            gas_fee_limit: None,
+            ttl: None,
+        };
+        expect_emitted_approve_event(
+            &rt,
+            sponsor_f4_eth_addr,
+            spender_f4_eth_addr,
+            approve_params.credit_limit.clone(),
+            approve_params.gas_fee_limit.clone(),
+            0,
+        );
+        let response = rt.call::<BlobsActor>(
+            Method::ApproveCredit as u64,
+            IpldBlock::serialize_cbor(&approve_params).unwrap(),
+        );
+        assert!(response.is_ok());
+        rt.verify();
+
+        // Try sending zero
+        rt.set_origin(spender_id_addr);
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, spender_id_addr);
+        rt.expect_validate_caller_any();
+        rt.set_received(TokenAmount::zero());
+        let hash = new_hash(1024);
+        let add_params = AddBlobParams {
+            from: spender_id_addr,
+            sponsor: Some(sponsor_id_addr),
+            hash: hash.0,
+            metadata_hash: new_hash(1024).0,
+            source: new_pk(),
+            id: SubscriptionId::default(),
+            size: hash.1,
+            ttl: Some(3600),
+        };
+        expect_get_config(&rt);
+        expect_emitted_add_event(&rt, 0, &add_params, sponsor_f4_eth_addr, add_params.size);
+        let response = rt.call::<BlobsActor>(
+            Method::AddBlob as u64,
+            IpldBlock::serialize_cbor(&add_params).unwrap(),
+        );
+        assert!(response.is_ok());
+        rt.verify();
+
+        // Try sending non-zero -> cannot buy for a sponsor, tokens are sent back
+        rt.set_origin(spender_id_addr);
+        rt.set_caller(*ETHACCOUNT_ACTOR_CODE_ID, spender_id_addr);
+        rt.expect_validate_caller_any();
+        let received = TokenAmount::from_whole(1);
+        rt.set_received(received.clone());
+        rt.set_balance(received.clone());
+        let hash = new_hash(1024);
+        let add_params = AddBlobParams {
+            from: spender_id_addr,
+            sponsor: Some(sponsor_id_addr),
+            hash: hash.0,
+            metadata_hash: new_hash(1024).0,
+            source: new_pk(),
+            id: SubscriptionId::default(),
+            size: hash.1,
+            ttl: Some(3600),
+        };
+        expect_get_config(&rt);
+        expect_emitted_add_event(&rt, 0, &add_params, sponsor_f4_eth_addr, add_params.size);
+        rt.expect_send_simple(
+            spender_id_addr,
+            METHOD_SEND,
+            None,
+            received,
+            None,
+            ExitCode::OK,
+        );
+        let response = rt.call::<BlobsActor>(
+            Method::AddBlob as u64,
+            IpldBlock::serialize_cbor(&add_params).unwrap(),
+        );
+        assert!(response.is_ok());
+        rt.verify();
+    }
+}
